This commit marks the official move to C++11 for jadebase.  Specific to this commit was the replacement of jade::shared_ptr with std::shared_ptr.

+++ src/windowsys/jb_windowevent.hpp
@@ -57,7 +57,7 @@ namespace jade
         float   position[ 2 ];                                                  // Position [ x, y ] relative to screen (fractional if supported)
         float   prev_pos[ 2 ];                                                  // Previous position [ x, y ], { NaN, NaN } if no previous
         float   pressure;                                                       // Pressure, ( 0.0 ... 1.0 )
-        int         tilt[ 2 ];                                                  // Tilt [ x, y ], ( -1.0 ... 1.0 )
+        float       tilt[ 2 ];                                                  // Tilt [ x, y ], ( -1.0 ... 1.0 )
         float   rotation;                                                       // Rotation (0.0 up to but not including 1.0 is a full rotation, can contain
                                                                                 // multiple rotations but usually not)
         float      wheel;                                                       // Tangential (wheel) pressure -1.0 through 1.0
